requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
python-dotenv==1.2.1
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
uritemplate==4.2.0
psycopg2==2.9.11

# Blockchain / Web3 integration
web3>=7.14.0

# IPFS integration
ipfshttpclient>=0.7.0

# For cryptographic operations
cryptography>=46.0.3

# For async task processing (optional, for production)
# celery>=5.3.0
# redis>=5.0.0
